- Delete Mixin for migration (P2P) - Add binging-check module
