+++ src/replication/trystero/P2PReplicatorCore.ts
@@ -1,14 +1,7 @@
 import type { SimpleStore } from "octagonal-wheels/databases/SimpleStoreBase";
 
 import { $msg } from "../../common/i18n";
-import {
-    LOG_LEVEL_INFO,
-    LOG_LEVEL_VERBOSE,
-    P2P_DEFAULT_SETTINGS,
-    type EntryDoc,
-    type LOG_LEVEL,
-    type P2PSyncSetting,
-} from "../../common/types";
+import { type EntryDoc, type LOG_LEVEL, type P2PSyncSetting } from "../../common/types";
 import { reactiveSource, type ReactiveSource } from "octagonal-wheels/dataobject/reactive";
 import { EVENT_DATABASE_REBUILT, EVENT_SETTING_SAVED } from "../../events/coreEvents";
 import { eventHub } from "../../hub/hub";
@@ -16,7 +9,7 @@ import type { Confirm } from "../../interfaces/Confirm";
 import { EVENT_PLATFORM_UNLOADED } from "../../PlatformAPIs/base/APIBase";
 import { setReplicatorFunc } from "./LiveSyncTrysteroReplicator";
 import type { CommandShim, PluginShim } from "./P2PReplicatorPaneCommon";
-import { TrysteroReplicator, type P2PReplicationProgress } from "./TrysteroReplicator";
+import { type P2PReplicationProgress } from "./TrysteroReplicator";
 import {
     EVENT_ADVERTISEMENT_RECEIVED,
     EVENT_DEVICE_LEAVED,
@@ -161,8 +154,6 @@ export class P2PLogCollector {
     p2pReplicationLine = reactiveSource("");
 }
 
-type Constructor<T> = new (...args: any[]) => T;
-
 export interface P2PReplicatorBase {
     storeP2PStatusLine: ReactiveSource<string>;
     plugin: PluginShim;
@@ -178,109 +169,4 @@ export interface P2PReplicatorBase {
     init(): Promise<this>;
 
     services: ServiceHub;
-}
-
-export function P2PReplicatorMixIn<TBase extends Constructor<P2PReplicatorBase>>(base: TBase) {
-    return class MixInP2PReplicatorCommand extends base implements CommandShim {
-        _replicatorInstance?: TrysteroReplicator;
-        p2pLogCollector = new P2PLogCollector();
-
-        afterConstructor() {
-            return;
-        }
-
-        getPlatform(): string {
-            return "unknown";
-        }
-
-        constructor(...args: any[]) {
-            super(...args);
-            setReplicatorFunc(() => this._replicatorInstance);
-            addP2PEventHandlers(this);
-            this.afterConstructor();
-        }
-        async open() {
-            await openP2PReplicator(this);
-        }
-        async close() {
-            await closeP2PReplicator(this);
-        }
-
-        getConfig(key: string) {
-            const vaultName = this.services.vault.getVaultName();
-            const dbKey = `${vaultName}-${key}`;
-            return localStorage.getItem(dbKey);
-        }
-        setConfig(key: string, value: string) {
-            const vaultName = this.services.vault.getVaultName();
-            const dbKey = `${vaultName}-${key}`;
-            localStorage.setItem(dbKey, value);
-        }
-        enableBroadcastCastings() {
-            return this?._replicatorInstance?.enableBroadcastChanges();
-        }
-        disableBroadcastCastings() {
-            return this?._replicatorInstance?.disableBroadcastChanges();
-        }
-
-        override init(): Promise<this> {
-            return super.init();
-        }
-
-        async initialiseP2PReplicator(): Promise<TrysteroReplicator> {
-            await this.init();
-            try {
-                if (this._replicatorInstance) {
-                    await this._replicatorInstance.close();
-                    this._replicatorInstance = undefined;
-                }
-
-                if (!this.settings.P2P_AppID) {
-                    this.settings.P2P_AppID = P2P_DEFAULT_SETTINGS.P2P_AppID;
-                }
-                const getInitialDeviceName = () =>
-                    this.getConfig("p2p_device_name") || this.services.vault.getVaultName();
-
-                const getSettings = () => this.settings;
-                const store = () => this.simpleStore();
-                const getDB = () => this.getDB();
-
-                const getConfirm = () => this.confirm;
-                const getPlatform = () => this.getPlatform();
-                const env = {
-                    get db() {
-                        return getDB();
-                    },
-                    get confirm() {
-                        return getConfirm();
-                    },
-                    get deviceName() {
-                        return getInitialDeviceName();
-                    },
-                    get platform() {
-                        return getPlatform();
-                    },
-                    get settings() {
-                        return getSettings();
-                    },
-                    processReplicatedDocs: async (docs: EntryDoc[]): Promise<void> => {
-                        await this.handleReplicatedDocuments(docs);
-                        // No op. This is a client and does not need to process the docs
-                    },
-                    get simpleStore() {
-                        return store();
-                    },
-                };
-                this._replicatorInstance = new TrysteroReplicator(env);
-                return this._replicatorInstance;
-            } catch (e) {
-                this._log(
-                    e instanceof Error ? e.message : "Something occurred on Initialising P2P Replicator",
-                    LOG_LEVEL_INFO
-                );
-                this._log(e, LOG_LEVEL_VERBOSE);
-                throw e;
-            }
-        }
-    };
 }
